// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        POOut: {
            /** Id */
            id: number;
            /** Po Number */
            po_number: string;
            /** Client Id */
            client_id: number;
            /**
             * Po Date
             * Format: date
             */
            po_date: string;
            /** Po Value */
            po_value: string;
            /** Agreed Payment Terms Days */
            agreed_payment_terms_days: number;
            /** Po Document Url */
            po_document_url: string | null;
            /**
             * Status
             * @enum {string}
             */
            status: "open" | "locked" | "closed";
            /** Is Allocation Balanced */
            is_allocation_balanced: boolean;
            /** Allocations */
            allocations: components["schemas"]["AllocationOut"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}